+++ docs/changelog.rst
@@ -27,6 +27,8 @@ Modifications to existing flavors:
   (`gh-242` <https://github.com/django/django-localflavor/pull/242>`_).
 - Fixed French FRNationalIdentificationNumber bug with corsican people born after 2000.
   (`gh-242` <https://github.com/django/django-localflavor/pull/242>`_).
+- Fixed the translation for US state 'Georgia' from colliding with the country 'Georgia'
+  (`gh-250` <https://github.com/django/django-localflavor/pull/250>`_).
 
 1.3   (2016-05-06)
 ------------------